add methods to archived a projet, stay in DB but no showing in main, addr outing to this new method
