+++ src/Controller/MainController.php
@@ -9,14 +9,5 @@ use Symfony\Component\Routing\Attribute\Route;
 
 final class MainController extends AbstractController
 {
-    #[Route('/', name: 'app_main', methods:['GET'])]
-    public function index(ProjectRepository $projectRepository): Response
-    {
-        $projects = $projectRepository->findAll();
-
-        return $this->render('main/index.html.twig', [
-            'controller_name' => 'MainController',
-            'projects' => $projects
-        ]);
-    }
+    
 }
