+++ templates/base.html.twig
@@ -23,7 +23,7 @@
                 </a>
                 <ul id="main-menu">
                 <li class="active">
-                    <a href="#">
+                    <a href="{{ path('app_main') }}">
                     <i class="fa-solid fa-list-check"></i> Projets
                     </a>
                 </li>
