+++ templates/main/index.html.twig
@@ -20,7 +20,7 @@
                             </a>
                             <div class="card-project-icons">
                                 <a href="{{ path('app_project_edit', {id: project.id})}}"><i class="fa-solid fa-pen"></i></a>
-                                <a href="#"><i class="fa-solid fa-trash"></i></a>
+                                <a href="{{ path('app_project_archive', {id: project.id})}}"><i class="fa-solid fa-trash"></i></a>
                             </div>
                         </article>  
                     {% endif %}
